server/pyproject.toml
[project]
name = "feedmeinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.127.0",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
